Renamed files and moved tests to better reflect the situation.
