+++ tests/test_config_parsing_old.py
@@ -1,18 +1,16 @@
-import copy
-
 import pytest
 
 from just_bin_it.histograms.histogram_factory import parse_config
 
 CONFIG_FULL = {
     "cmd": "config",
+    "data_brokers": ["localhost:9092"],
+    "data_topics": ["junk_data_2"],
     "start": 1571831082207,
     "stop": 1571831125940,
     "histograms": [
         {
             "type": "hist1d",
-            "data_brokers": ["localhost:9092"],
-            "data_topics": ["junk_data_2"],
             "tof_range": [0, 100000000],
             "det_range": [0, 100],
             "num_bins": 50,
@@ -21,8 +19,6 @@ CONFIG_FULL = {
         },
         {
             "type": "hist1d",
-            "data_brokers": ["localhost:9092"],
-            "data_topics": ["junk_data_2"],
             "tof_range": [0, 100000000],
             "det_range": [0, 100],
             "num_bins": 50,
@@ -33,11 +29,10 @@ CONFIG_FULL = {
 }
 
 
-class TestConfigParserV2:
+class TestConfigParserOldStyle:
     """
-    Tests specific to the new style configuration messages.
-    Eventually can be merged in the original tests when the old-style syntax is
-    retired.
+    Tests specific to the old style configuration messages.
+    Can be deleted when the old-style syntax is retired.
     """
 
     @pytest.fixture(autouse=True)
@@ -52,25 +47,11 @@ class TestConfigParserV2:
         assert stop == CONFIG_FULL["stop"]
 
         for i, h in enumerate(hists):
-            assert h["data_brokers"] == CONFIG_FULL["histograms"][i]["data_brokers"]
-            assert h["data_topics"] == CONFIG_FULL["histograms"][i]["data_topics"]
+            assert h["data_brokers"] == CONFIG_FULL["data_brokers"]
+            assert h["data_topics"] == CONFIG_FULL["data_topics"]
             assert h["type"] == CONFIG_FULL["histograms"][i]["type"]
             assert h["tof_range"] == CONFIG_FULL["histograms"][i]["tof_range"]
             assert h["det_range"] == CONFIG_FULL["histograms"][i]["det_range"]
             assert h["num_bins"] == CONFIG_FULL["histograms"][i]["num_bins"]
             assert h["topic"] == CONFIG_FULL["histograms"][i]["topic"]
             assert h["id"] == CONFIG_FULL["histograms"][i]["id"]
-
-    def test_raises_if_no_data_brokers_supplied_for_a_hist(self):
-        config = copy.deepcopy(CONFIG_FULL)
-        del config["histograms"][0]["data_brokers"]
-
-        with pytest.raises(Exception):
-            parse_config(config)
-
-    def test_raises_if_no_data_topics_supplied_for_a_hist(self):
-        config = copy.deepcopy(CONFIG_FULL)
-        del config["histograms"][0]["data_topics"]
-
-        with pytest.raises(Exception):
-            parse_config(config)
